[FIX] Fix tournament router bug
In commit '23a542cd01 ('[UPDATE] Passing game data in uri is bad')',
changed method passing game data, from URI to radio.

But it caused a problem. When you win a tournament round,
it redirects to 'play', in route 'play'. routing to same
route does nothing in Backbone. (It was okay before, because
the route was 'play?channelId=XXXX....'. it was unique)

Fix it by routing to loading page when you win a round