[FIX] Refresh only when in ChatCollectionView is rendered
ChatCollectionView is refreshing when the current tab is not
chat tab. Tried to fix it before, but it turned out to be bad solution.
So that commit was reverted...

This commit will make fetch radio of ChatCollectionView will be
destroyed on onDestroy function, will be made in onRender function.

v2:
      Tried to put creation of radio on onInitialize function,
      but when refreshing tab:
	    [1] a new ChatCollectionView is created, Radio created
	    [2] an old ChatCollectionView is destroyed, Radio destroyed

      Radio is global, so it didn't work as expected.

      So make radio on onRender function.